Vencord Standalone without git/node
